Separate cached / from web calls, add components folder
