+++ src/ui/components/search-repo-form.js
@@ -1,5 +1,5 @@
 import { h } from 'preact';
-import { useContext, useState } from 'preact/hooks';
+import { useContext } from 'preact/hooks';
 import { store, ACTIONS  } from 'ui/store';
 
 const handleInput = (dispatch, action) => event =>
