+++ src/ui/index.js
@@ -1,7 +1,7 @@
 import {h, render } from 'preact';
 import StateProvider from 'ui/store';
-import RadioGroup from 'ui/radio-group';
-import RepoFormDispatcher from 'ui/repo-form-dispatcher';
+import RadioGroup from 'ui/components/radio-group';
+import RepoFormDispatcher from 'ui/components/repo-form-dispatcher';
 
 
 const App = () => (
